+++ apis/v2/about/future-careers/index.ts
@@ -1,13 +1,4 @@
-import { getRequest } from '@/apis';
 import { putRequest } from '@/apis';
-import { FutureCareers } from '@/apis/types/about';
-import { FETCH_TAG_CAREER } from '@/constants/network';
-import { Language } from '@/types/language';
-
-export const getFutureCareeres = (language: Language) =>
-  getRequest<FutureCareers>(`/v2/about/future-careers?language=${language}`, undefined, {
-    next: { tags: [FETCH_TAG_CAREER] },
-  });
 
 export const putFutureCareers = (data: { koDescription: string; enDescription: string }) =>
   putRequest('/v2/about/future-careers', {
