+++ app/[locale]/about/history/page.tsx
@@ -1,6 +1,6 @@
 export const dynamic = 'force-dynamic';
 
-import { getHistory } from '@/apis/v2/about/history';
+import { getHistory } from '@/apis/v1/about/history';
 import { EditButton } from '@/components/common/Buttons';
 import LoginVisible from '@/components/common/LoginVisible';
 import HTMLViewer from '@/components/form/html/HTMLViewer';
