+++ app/[locale]/about/overview/edit/page.tsx
@@ -1,5 +1,5 @@
 import { putOverviewAction } from '@/actions/about';
-import { getOverview } from '@/apis/v2/about/overview';
+import { getOverview } from '@/apis/v1/about/overview';
 import PageLayout from '@/components/layout/pageLayout/PageLayout';
 import { overview } from '@/constants/segmentNode';
 import { getEditorFile, getEditorImage } from '@/utils/formData';
